Make sure the architecture define is always set
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@7026 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/Makensisw/toolbar.cpp
@@ -196,7 +196,7 @@ static UINT GetToolbarDropdownMenuPos(HWND hTB, UINT Id, POINT&pt)
 static void ShowToolbarDropdownMenu(const NMTOOLBAR&nmtb, HWND hNotifyWnd, HMENU hParentMenu, UINT SubMenuId = -1)
 {
   POINT pt;
-  HMENU hMenu = SubMenuId == -1 ? hParentMenu : FindSubMenu(hParentMenu, SubMenuId);
+  HMENU hMenu = SubMenuId == static_cast<UINT>(-1) ? hParentMenu : FindSubMenu(hParentMenu, SubMenuId);
   UINT tpmf = GetToolbarDropdownMenuPos(nmtb.hdr.hwndFrom, nmtb.iItem, pt);
   TrackPopupMenu(hMenu, tpmf, pt.x, pt.y, 0, hNotifyWnd, NULL);
 }
